ada-driving-assistant/requirements.txt
# Driving assistant dashboard
anthropic>=0.40.0
flask>=3.0.0
python-dotenv>=1.0.0

# AWS / simulation
boto3>=1.35.0
